requirements.txt
# Core deep learning
tensorflow==2.12.0
keras==2.12.0

# Data & science
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.10.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Config & utilities
PyYAML>=6.0
tqdm>=4.65.0
joblib>=1.2.0

# Optional: Kaggle dataset download
kaggle>=1.5.13

# Notebook support
jupyter>=1.0.0
ipykernel>=6.22.0
